requirements.txt
flask
scikit-learn
pandas
pyzbar
Pillow
